refactor: move platform files
