+++ src/MeoAssistant/Controller.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #ifdef _WIN32
-#include "Utils/SafeWindows.h"
+#include "Utils/Platform/SafeWindows.h"
 #include <mswsock.h>
 #endif
