+++ src/MeoAssistant/Utils/Platform/AsstPlatformPosix.cpp
@@ -1,7 +1,7 @@
 #if __has_include(<unistd.h>)
 #include "AsstPlatformPosix.h"
 #include "AsstPlatform.h"
-#include "AsstUtils.hpp"
+#include "Utils/AsstUtils.hpp"
 
 #include <cstdlib>
