+++ src/MeoAssistant/Utils/Platform/SafeWindows.h
